siamese_detector/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
opencv-python-headless>=4.9.0
scikit-image>=0.22.0
numpy>=1.26.0
Pillow>=10.0.0
pydantic>=2.0.0
